+++ src/AngleSharp.Io/Interfaces/ILocalStorage.cs
@@ -1,0 +1,9 @@
+namespace AngleSharp.Io.Dom
+{
+    /// <summary>
+    /// Represents a local storage implementation.
+    /// </summary>
+    public interface ILocalStorage : IStorage
+    {
+    }
+}