+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Released on Friday, July 31 2026.
 
 - Updated to use a minimum of AngleSharp 1.5
 - Fixed domain handling in cookies #36
+- Added `LocalStorage` and `SessionStorage`
 
 # 1.0.0